@@ -1,15 +1,13 @@
-import React, { Component } from "react";
+import React from "react";
 import "./App.css";
 import Hangman from "./Hangman/Hangman";
 
-class App extends Component {
-  render() {
-    return (
-      <div className="App">
-        <Hangman />
-      </div>
-    );
-  }
+function App() {
+  return (
+    <div className="App">
+      <Hangman />
+    </div>
+  );
 }
 
 export default App;
